+++ src/main/java/com/stockcontrol/RawMaterial/services/RawMaterialService.java
@@ -4,6 +4,7 @@ import com.stockcontrol.RawMaterial.enums.TypeProduct;
 import com.stockcontrol.RawMaterial.exceptions.BarcodeExistException;
 import com.stockcontrol.RawMaterial.exceptions.CodeExistException;
 import com.stockcontrol.RawMaterial.models.RawMaterialModel;
+import com.stockcontrol.RawMaterial.models.StockModel;
 import com.stockcontrol.RawMaterial.repositories.RawMaterialRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -17,6 +18,7 @@ import java.util.Optional;
 public class RawMaterialService {
 
     final RawMaterialRepository rawMaterialRepository;
+    final StockService stockService;
 
     @Transactional
     public RawMaterialModel save(RawMaterialModel RawMaterialModel){
@@ -30,12 +32,24 @@ public class RawMaterialService {
             throw new BarcodeExistException("The barcode already exists: " + RawMaterialModel.getBarcode() + ". Enter a new code.");
         }
         RawMaterialModel = rawMaterialRepository.save(RawMaterialModel);
+
+        StockModel stockModel = new StockModel();
+        stockModel.setCode(RawMaterialModel.getCode());
+        stockModel.setDescription(RawMaterialModel.getDescription());
+        stockModel.setTypeProduct(RawMaterialModel.getTypeProduct());
+        stockModel.setStockQuantity(0.00);
+        stockModel.setBarcode(RawMaterialModel.getBarcode());
+
+        stockService.createdStock(stockModel);
         return RawMaterialModel;
     }
 
     @Transactional
     public Optional<RawMaterialModel> delete(Optional<RawMaterialModel> product0){
-        rawMaterialRepository.delete(product0.get());
+        if (product0.isPresent()){
+            rawMaterialRepository.delete(product0.get());
+            stockService.delete(product0.get().getCode());
+        }
         return product0;
     }
